requirements-dev.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-allauth==0.42.0
django-cleanup==5.0.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
